[refactor] Restructure UI packages
