+++ core/src/space/pxls/renderers/Heatmap.java
@@ -5,16 +5,14 @@ import com.badlogic.gdx.Net;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Pixmap;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.TextureData;
 import com.badlogic.gdx.graphics.glutils.FrameBuffer;
-import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Vector2;
 
 import java.util.Timer;
 import java.util.TimerTask;
 
 import space.pxls.Pxls;
-import space.pxls.ui.CanvasScreen;
+import space.pxls.ui.Screens.CanvasScreen;
 
 public class Heatmap implements Renderer {
     public final Color fillStyle = Color.valueOf("#CD5C5C");
