+++ core/src/space/pxls/ui/Components/PxlsButton.java
@@ -1,11 +1,10 @@
-package space.pxls.ui;
+package space.pxls.ui.Components;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.utils.Align;
 
 import space.pxls.Pxls;
-import space.pxls.ui.Components.TTFLabel;
 
 public class PxlsButton extends TextButton {
     public PxlsButton(String text) {
