+++ core/src/space/pxls/ui/Components/PxlsCheckBox.java
@@ -1,19 +1,15 @@
-package space.pxls.ui;
+package space.pxls.ui.Components;
 
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
-import com.badlogic.gdx.scenes.scene2d.ui.Button;
 import com.badlogic.gdx.scenes.scene2d.ui.Cell;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
-import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-import com.badlogic.gdx.scenes.scene2d.utils.Disableable;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
 
 import space.pxls.Pxls;
-import space.pxls.ui.Components.TTFLabel;
 
 /**
  * A button with 2 states to represent checked/unchecked, and a label. Pretty simple.
